refactor: rename startpage module and route to dashboard
