+++ src/views/dashboard/components/BranchPhotoCard.vue
