+++ src/views/dashboard/components/DashboardBlock.vue
