+++ src/views/dashboard/components/DashboardTabbedBlock.vue
